+++ proj1/test/test_restore_10.sh
@@ -27,11 +27,11 @@ cd build
 rm -rf 1 2
 rm -rf testfiles
 mkdir testfiles
-curl http://ftp.debian.org/debian/dists/jessie/main/source/Release -o testfiles/source_Release  # 102B
-curl http://ftp.debian.org/debian/dists/jessie/Release             -o testfiles/Release         # 77.3KB
-curl http://ftp.debian.org/debian/dists/jessie/ChangeLog           -o testfiles/ChangeLog       # 2.3MB
-timeout $TIMEOUT java PeerDriver $VERSION 1 service1 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT   & PID1=$!
-timeout $TIMEOUT java PeerDriver 1.0 2 service2 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID2=$!
+if ! [ -f testfiles/source_Release ]; then curl http://ftp.debian.org/debian/dists/jessie/main/source/Release -o testfiles/source_Release; fi # 102B
+if ! [ -f testfiles/Release        ]; then curl http://ftp.debian.org/debian/dists/jessie/Release             -o testfiles/Release       ; fi # 77.3KB
+if ! [ -f testfiles/ChangeLog      ]; then curl http://ftp.debian.org/debian/dists/jessie/ChangeLog           -o testfiles/ChangeLog     ; fi # 2.3MB
+timeout $TIMEOUT java PeerDriver $VERSION 1 service1 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID1=$!
+timeout $TIMEOUT java PeerDriver 1.0      2 service2 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID2=$!
 echo "Started peers with PIDs $PID1, $PID2"
 sleep 1
